ml_server/requirements.txt
fastapi==0.96.1
openai==0.27.8
pydantic==1.9.0
pymongo==4.3.3
python-dotenv
uvicorn
